+++ lib/widgets/audio/audio_player.dart
@@ -1,10 +1,14 @@
+import 'dart:io';
+
 import 'package:audio_video_progress_bar/audio_video_progress_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lotti/blocs/audio/player_cubit.dart';
 import 'package:lotti/blocs/audio/player_state.dart';
+import 'package:lotti/blocs/journal/entry_cubit.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/themes/theme.dart';
+import 'package:lotti/widgets/journal/entry_tools.dart';
 
 class AudioPlayerWidget extends StatelessWidget {
   const AudioPlayerWidget(this.journalAudio, {super.key});
@@ -36,6 +40,9 @@ class AudioPlayerWidget extends StatelessWidget {
     return BlocBuilder<AudioPlayerCubit, AudioPlayerState>(
       builder: (BuildContext context, AudioPlayerState state) {
         final isActive = state.audioNote?.meta.id == journalAudio.meta.id;
+        final cubit = context.read<AudioPlayerCubit>();
+        final entryCubit = context.read<EntryCubit>();
+        final transcripts = journalAudio.data.transcripts;
 
         return Column(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -52,8 +59,9 @@ class AudioPlayerWidget extends StatelessWidget {
                       ? styleConfig().activeAudioControl
                       : styleConfig().secondaryTextColor,
                   onPressed: () {
-                    context.read<AudioPlayerCubit>().setAudioNote(journalAudio);
-                    context.read<AudioPlayerCubit>().play();
+                    cubit
+                      ..setAudioNote(journalAudio)
+                      ..play();
                   },
                 ),
                 IgnorePointer(
@@ -67,30 +75,28 @@ class AudioPlayerWidget extends StatelessWidget {
                         iconSize: 32,
                         tooltip: 'Rewind 15s',
                         color: styleConfig().secondaryTextColor,
-                        onPressed: () => context.read<AudioPlayerCubit>().rew(),
+                        onPressed: cubit.rew,
                       ),
                       IconButton(
                         icon: const Icon(Icons.pause),
                         iconSize: 32,
                         tooltip: 'Pause',
                         color: styleConfig().secondaryTextColor,
-                        onPressed: () =>
-                            context.read<AudioPlayerCubit>().pause(),
+                        onPressed: cubit.pause,
                       ),
                       IconButton(
                         icon: const Icon(Icons.fast_forward),
                         iconSize: 32,
                         tooltip: 'Fast forward 15s',
                         color: styleConfig().secondaryTextColor,
-                        onPressed: () => context.read<AudioPlayerCubit>().fwd(),
+                        onPressed: cubit.fwd,
                       ),
                       IconButton(
                         icon: const Icon(Icons.stop),
                         iconSize: 32,
                         tooltip: 'Stop',
                         color: styleConfig().secondaryTextColor,
-                        onPressed: () =>
-                            context.read<AudioPlayerCubit>().stopPlay(),
+                        onPressed: cubit.stopPlay,
                       ),
                       IconButton(
                         icon: Text(
@@ -105,13 +111,29 @@ class AudioPlayerWidget extends StatelessWidget {
                         ),
                         iconSize: 32,
                         tooltip: 'Toggle speed',
-                        onPressed: () => context
-                            .read<AudioPlayerCubit>()
-                            .setSpeed(speedToggleMap[state.speed] ?? 1),
+                        onPressed: () =>
+                            cubit.setSpeed(speedToggleMap[state.speed] ?? 1),
                       ),
                     ],
                   ),
                 ),
+                if (Platform.isMacOS)
+                  IconButton(
+                    icon: const Icon(Icons.transcribe_outlined),
+                    iconSize: 20,
+                    tooltip: 'Transcribe',
+                    color: styleConfig().secondaryTextColor,
+                    onPressed: () async {
+                      await cubit.setAudioNote(journalAudio);
+                      await cubit.transcribe();
+                      await Future<void>.delayed(
+                        const Duration(milliseconds: 100),
+                      );
+                      entryCubit
+                        ..setController()
+                        ..emitState();
+                    },
+                  ),
               ],
             ),
             Row(
@@ -128,9 +150,7 @@ class AudioPlayerWidget extends StatelessWidget {
                     thumbColor: Colors.white,
                     barHeight: 3,
                     thumbRadius: 5,
-                    onSeek: (newPosition) {
-                      context.read<AudioPlayerCubit>().seek(newPosition);
-                    },
+                    onSeek: cubit.seek,
                     timeLabelTextStyle: monospaceTextStyle().copyWith(
                       color: styleConfig().secondaryTextColor,
                     ),
@@ -138,9 +158,94 @@ class AudioPlayerWidget extends StatelessWidget {
                 ),
               ],
             ),
+            if (transcripts?.isNotEmpty ?? false)
+              Column(
+                children: [
+                  const SizedBox(height: 10),
+                  ...transcripts!.map(
+                    TranscriptListItem.new,
+                  ),
+                ],
+              ),
           ],
         );
       },
+    );
+  }
+}
+
+class TranscriptListItem extends StatefulWidget {
+  const TranscriptListItem(
+    this.transcript, {
+    super.key,
+  });
+
+  final AudioTranscript transcript;
+
+  @override
+  State<TranscriptListItem> createState() => _TranscriptListItemState();
+}
+
+class _TranscriptListItemState extends State<TranscriptListItem> {
+  bool show = false;
+
+  void toggleShow() {
+    setState(() {
+      show = !show;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(
+        vertical: 2,
+        horizontal: 30,
+      ),
+      child: Column(
+        children: [
+          MouseRegion(
+            cursor: SystemMouseCursors.click,
+            child: GestureDetector(
+              onTap: toggleShow,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Text(
+                    '${dfShorter.format(widget.transcript.created)}  '
+                    '${formatSeconds(widget.transcript.processingTime)}  '
+                    'Language: ${widget.transcript.detectedLanguage}    ',
+                    style: transcriptHeaderStyle(),
+                  ),
+                  Text(
+                    '${widget.transcript.library}, '
+                    ' ${widget.transcript.model}',
+                    style: transcriptHeaderStyle(),
+                  ),
+                  if (show)
+                    const Icon(
+                      Icons.keyboard_double_arrow_up_outlined,
+                      size: 15,
+                    )
+                  else
+                    const Icon(
+                      Icons.keyboard_double_arrow_down_outlined,
+                      size: 15,
+                    ),
+                ],
+              ),
+            ),
+          ),
+          if (show)
+            Padding(
+              padding: const EdgeInsets.only(bottom: 10),
+              child: SelectableText(
+                widget.transcript.transcript,
+                style: transcriptStyle(),
+              ),
+            ),
+        ],
+      ),
     );
   }
 }