+++ macos/Runner/MainFlutterWindow.swift
@@ -1,15 +1,42 @@
 import Cocoa
 import FlutterMacOS
 
+import IOKit
+
 class MainFlutterWindow: NSWindow {
-  override func awakeFromNib() {
-    let flutterViewController = FlutterViewController.init()
-    let windowFrame = self.frame
-    self.contentViewController = flutterViewController
-    self.setFrame(windowFrame, display: true)
-
-    RegisterGeneratedPlugins(registry: flutterViewController)
-
-    super.awakeFromNib()
-  }
+    override func awakeFromNib() {
+        let flutterViewController = FlutterViewController.init()
+        let windowFrame = self.frame
+        self.contentViewController = flutterViewController
+        self.setFrame(windowFrame, display: true)
+        
+        let transcriptionChannel = FlutterMethodChannel(
+            name: "lotti/transcribe",
+            binaryMessenger: flutterViewController.engine.binaryMessenger)
+        
+        transcriptionChannel.setMethodCallHandler { (call, result) in
+            switch call.method {
+            case "transcribe":
+                guard let args = call.arguments as? [String: Any] else { return }
+                let audioFilePath = args["audioFilePath"] as! String
+                let modelPath = args["modelPath"] as! String
+                Task {
+                   await transcribe(audioFilePath: audioFilePath, modelPath: modelPath, result: result)
+                }
+            case "detectLanguage":
+                guard let args = call.arguments as? [String: Any] else { return }
+                let audioFilePath = args["audioFilePath"] as! String
+                let modelPath = args["modelPath"] as! String
+                Task {
+                   await detectLanguage(audioFilePath: audioFilePath, modelPath: modelPath, result: result)
+                }
+            default:
+                result(FlutterMethodNotImplemented)
+            }
+        }
+        
+        RegisterGeneratedPlugins(registry: flutterViewController)
+        
+        super.awakeFromNib()
+    }
 }