+++ test/widgets/audio/audio_player_test.dart
@@ -3,6 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/blocs/audio/player_cubit.dart';
 import 'package:lotti/blocs/audio/player_state.dart';
+import 'package:lotti/blocs/journal/entry_cubit.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/audio/audio_player.dart';
@@ -15,6 +16,8 @@ import '../../widget_test_utils.dart';
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
   registerFallbackValue(FakeJournalAudio());
+
+  final entryCubit = MockEntryCubit();
 
   group('AudioPlayerWidget Widget Tests - ', () {
     setUp(() {
@@ -55,10 +58,13 @@ void main() {
 
       await tester.pumpWidget(
         makeTestableWidgetWithScaffold(
-          BlocProvider<AudioPlayerCubit>(
-            create: (_) => mockAudioPlayerCubit,
-            lazy: false,
-            child: AudioPlayerWidget(pausedState.audioNote!),
+          BlocProvider<EntryCubit>.value(
+            value: entryCubit,
+            child: BlocProvider<AudioPlayerCubit>(
+              create: (_) => mockAudioPlayerCubit,
+              lazy: false,
+              child: AudioPlayerWidget(pausedState.audioNote!),
+            ),
           ),
         ),
       );
@@ -119,10 +125,13 @@ void main() {
 
       await tester.pumpWidget(
         makeTestableWidgetWithScaffold(
-          BlocProvider<AudioPlayerCubit>(
-            create: (_) => mockAudioPlayerCubit,
-            lazy: false,
-            child: AudioPlayerWidget(playingState.audioNote!),
+          BlocProvider<EntryCubit>.value(
+            value: entryCubit,
+            child: BlocProvider<AudioPlayerCubit>(
+              create: (_) => mockAudioPlayerCubit,
+              lazy: false,
+              child: AudioPlayerWidget(playingState.audioNote!),
+            ),
           ),
         ),
       );
@@ -161,6 +170,73 @@ void main() {
 
       await tester.tap(stopIconFinder);
       verify(mockAudioPlayerCubit.stopPlay).called(1);
+    });
+
+    testWidgets('controls are are displayed, playing state', (tester) async {
+      final playingState = AudioPlayerState(
+        status: AudioPlayerStatus.playing,
+        progress: const Duration(seconds: 15),
+        totalDuration: const Duration(minutes: 1),
+        pausedAt: Duration.zero,
+        speed: 1,
+        audioNote: testAudioEntryWithTranscripts,
+      );
+
+      when(() => mockAudioPlayerCubit.stream).thenAnswer(
+        (_) => Stream<AudioPlayerState>.fromIterable([playingState]),
+      );
+
+      when(() => mockAudioPlayerCubit.state).thenAnswer(
+        (_) => playingState,
+      );
+
+      when(() => mockAudioPlayerCubit.setAudioNote(any()))
+          .thenAnswer((_) async {});
+
+      when(mockAudioPlayerCubit.close).thenAnswer((_) async {});
+      when(mockAudioPlayerCubit.stopPlay).thenAnswer((_) async {});
+      when(mockAudioPlayerCubit.pause).thenAnswer((_) async {});
+
+      when(() => mockAudioPlayerCubit.setSpeed(1.25)).thenAnswer((_) async {});
+
+      await tester.pumpWidget(
+        makeTestableWidgetWithScaffold(
+          BlocProvider<EntryCubit>.value(
+            value: entryCubit,
+            child: BlocProvider<AudioPlayerCubit>(
+              create: (_) => mockAudioPlayerCubit,
+              lazy: false,
+              child: SingleChildScrollView(
+                scrollDirection: Axis.horizontal,
+                child: AudioPlayerWidget(playingState.audioNote!),
+              ),
+            ),
+          ),
+        ),
+      );
+
+      await tester.pumpAndSettle();
+
+      final showIcon = find.byIcon(Icons.keyboard_double_arrow_down_outlined);
+      final hideIcon = find.byIcon(Icons.keyboard_double_arrow_up_outlined);
+
+      expect(find.text('whisper-v1.4.0,  ggml-small.bin'), findsOneWidget);
+
+      await tester.pumpAndSettle();
+
+      expect(find.text('transcript'), findsNothing);
+
+      expect(showIcon, findsOneWidget);
+
+      await tester.tap(showIcon);
+      await tester.pumpAndSettle();
+
+      expect(find.text('transcript'), findsOneWidget);
+
+      await tester.tap(hideIcon);
+      await tester.pumpAndSettle();
+
+      expect(find.text('transcript'), findsNothing);
     });
   });
 }