+++ test/widgets/journal/entry_details/entry_datetime_widget_test.dart
@@ -30,6 +30,7 @@ void main() {
           entry: testTextEntry,
           showMap: false,
           isFocused: false,
+          epoch: 0,
         ),
       );
     });